requirements.txt
bing_image_downloader==1.1.2
datasets==2.14.6
faiss==1.7.4
fasttext==0.9.2
google_images_search==1.4.6
hydra-core==1.3.2
img2vec_pytorch==1.0.1
matplotlib==3.8.3
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
pandas==2.2.1
Pillow==10.0.1
plotly==5.9.0
Requests==2.31.0
scikit_learn==1.4.0
scipy==1.12.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
wandb==0.16.2
